Move error helper into dedicated package
