+++ controllers/update_pod_config.go
@@ -25,6 +25,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/FoundationDB/fdb-kubernetes-operator/v2/internal/errors"
 	"github.com/apple/foundationdb/fdbkubernetesmonitor/api"
 
 	"github.com/go-logr/logr"
@@ -126,7 +127,7 @@ func (updatePodConfig) reconcile(
 				curLogger.Error(err, "Update Pod ConfigMap annotation")
 			}
 
-			if internal.IsNetworkError(err) &&
+			if errors.IsNetworkError(err) &&
 				processGroup.GetConditionTime(fdbv1beta2.SidecarUnreachable) == nil {
 				curLogger.Info("process group sidecar is not reachable")
 				processGroup.UpdateCondition(fdbv1beta2.SidecarUnreachable, true)
