+++ internal/errors/error_helper_test.go
@@ -18,7 +18,7 @@
  * limitations under the License.
  */
 
-package internal
+package errors
 
 import (
 	"fmt"
